+++ rasa/shared/importers/multi_project.py
@@ -55,6 +55,10 @@ class MultiProjectImporter(TrainingDataImporter):
 
         mark_as_experimental_feature(feature_name="MultiProjectImporter")
 
+    def get_config_file_for_auto_config(self) -> Optional[Text]:
+        """Returns config file path for auto-config only if there is a single one."""
+        return None
+
     def _init_from_path(self, path: Text) -> None:
         if os.path.isfile(path):
             self._init_from_file(path)
